# server/docs/swagger.yaml
definitions:
  api.ClerkCallbackErrorResponse:
    properties:
      error:
        example: Missing svix headers
        type: string
    type: object
  api.UserProfileData:
    properties:
      email:
        example: a***e@example.com
        type: string
      kycChecked:
        example: true
        type: boolean
      registeredAt:
        example: "2026-05-16T10:00:00Z"
        type: string
      username:
        example: alice
        type: string
    type: object
  api.UserProfileHTTPResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/api.UserProfileData'
      message:
        example: success
        type: string
    type: object
  data.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      err_msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /identity-ms/v1/clerk/callback:
    post:
      consumes:
      - application/json
      description: Verifies Svix signature headers and accepts Clerk webhook events
        (e.g. user.created).
      parameters:
      - description: Svix message ID
        in: header
        name: svix-id
        required: true
        type: string
      - description: Svix HMAC signature
        in: header
        name: svix-signature
        required: true
        type: string
      - description: Unix timestamp of the webhook message
        in: header
        name: svix-timestamp
        required: true
        type: string
      - description: Raw Clerk webhook event JSON payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: webhook accepted
        "400":
          description: missing svix headers, invalid body, or invalid signature
          schema:
            $ref: '#/definitions/api.ClerkCallbackErrorResponse'
        "500":
          description: webhook verification init failed
          schema:
            $ref: '#/definitions/api.ClerkCallbackErrorResponse'
      summary: Clerk webhook callback
      tags:
      - clerk-webhook
  /identity-ms/v1/kyc/singpass/callback:
    get:
      description: Validates OAuth state and completes KYC for the user who started
        login.
      parameters:
      - description: Singpass code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: callback accepted
        "400":
          description: invalid code or state
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Singpass callback
      tags:
      - singpass
  /identity-ms/v1/web/kyc/singpass/login:
    get:
      description: Requires authentication. Returns an authorize URL; the client should
        redirect the browser to it.
      produces:
      - application/json
      responses:
        "200":
          description: authorize url
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "401":
          description: authentication required
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Start Singpass KYC
      tags:
      - singpass
  /identity-ms/v1/web/user-profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/api.UserProfileHTTPResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/data.BaseResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/data.BaseResponse'
      summary: Get user profile (user)
      tags:
      - user-profile
swagger: "2.0"